requirements.txt
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
psycopg==3.2.10
psycopg-binary==3.2.10
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
